+++ test/support/slack/commands/get-sites.test.js
@@ -80,6 +80,9 @@ describe('GetSitesCommand', () => {
       Site: {
         allWithLatestAudit: sinon.stub(),
       },
+      Organization: {
+        findByImsOrgId: sinon.stub().resolves(),
+      },
     };
     boltAppStub = {
       action: sinon.stub(),
@@ -182,6 +185,30 @@ describe('GetSitesCommand', () => {
       const command = GetSitesCommand(context);
 
       await command.handleExecution(['other'], slackContext);
+
+      expect(slackContext.say.called).to.be.true;
+    });
+
+    it('handles command execution with non existing ims org id', async () => {
+      dataAccessStub.Site.allWithLatestAudit.resolves(generateSites(3));
+      const command = GetSitesCommand(context);
+
+      await command.handleExecution(['ASASASASASASASASASASASAS@AdobeOrg'], slackContext);
+
+      expect(slackContext.say.called).to.be.true;
+    });
+
+    it('handles command execution with ims org id', async () => {
+      const sites = generateSites(3);
+      const org = {
+        getId: () => 'some-org-id',
+        getSites: () => sites,
+      };
+      dataAccessStub.Site.allWithLatestAudit.resolves(sites);
+      dataAccessStub.Organization.findByImsOrgId.resolves(org);
+      const command = GetSitesCommand(context);
+
+      await command.handleExecution(['ASASASASASASASASASASASAS@AdobeOrg'], slackContext);
 
       expect(slackContext.say.called).to.be.true;
     });